Add MapsActivity for displaying a Google Map
This commit adds a new activity, MapsActivity, to the project. The activity
displays a Google Map using the Google Maps API and sets the initial location
of the map to the EPFL campus in Switzerland. A marker is also added to the
map at a specific location with a label. Finally, a listener is added to the
marker's info window click event to display a Toast message with the
coordinates of the clicked marker.