Add Dice Roll Feature
Could be much more, but I think that we should fix the technical debt before adding more stuff on top of an already bloated file.
The missing features would include making sure that all 3 dice rolls yield 3 different locations, as well as detecting when the player reaches the target location.
Another issue is that Location class name is not well suited for this class, as it conflicts with the built-in Location type that is used for the Map. Property would be a better suited name.
The statements above will naturally be turned into issues on Github.